pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "CarlosEduardoCardozo", email = "CarlosEduardoCardozo"},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "setuptools>=68.2.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]
